weaviate_local/requirements.txt
PyMuPDF==1.25.2
weaviate-client==4.15.4
